requirements.txt
aiohttp==3.11.14
fire==0.7.0
loguru==0.7.3
PyYAML==6.0.2
GitPython==3.1.44
